Split files out, correct error handling, tests passing
