+++ saml.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"net/url"
-	"strconv"
 	"time"
 
 	"github.com/beevik/etree"
@@ -26,14 +25,11 @@ type SAMLServiceProvider struct {
 	IDPCertificateStore         dsig.X509CertificateStore
 	SPKeyStore                  dsig.X509KeyStore
 	NameIdFormat                string
+	SkipSignatureValidation     bool
 }
 
 func (sp *SAMLServiceProvider) SigningContext() *dsig.SigningContext {
 	return dsig.NewDefaultSigningContext(sp.SPKeyStore)
-}
-
-func (sp *SAMLServiceProvider) validationContext() *dsig.ValidationContext {
-	return dsig.NewDefaultValidationContext(sp.IDPCertificateStore)
 }
 
 func (sp *SAMLServiceProvider) BuildAuthRequest() (string, error) {
@@ -52,6 +48,9 @@ func (sp *SAMLServiceProvider) BuildAuthRequest() (string, error) {
 	authnRequest.CreateAttr("AssertionConsumerServiceIndex", "0")
 	authnRequest.CreateAttr("AttributeConsumingServiceIndex", "0")
 	authnRequest.CreateAttr("IssueInstant", time.Now().UTC().Format(issueInstantFormat))
+
+	authnRequest.CreateAttr("Destination", "http://idp.astuart.co/idp/profile/SAML2/Redirect/SSO")
+	// authnRequest.CreateAttr("Destination", sp.IdentityProviderSSOURL)
 
 	authnRequest.CreateElement("saml:Issuer").SetText(sp.IdentityProviderIssuer)
 
@@ -154,7 +153,7 @@ func (sp *SAMLServiceProvider) Validate(el *etree.Element) error {
 
 	assertionElement := el.FindElement(AssertionTag)
 	if assertionElement == nil {
-		return errors.New("Missing Assertion element")
+		return ErrMissingAssertion
 	}
 
 	subjectStatement := assertionElement.FindElement(childPath(assertionElement.Space, SubjectTag))
@@ -197,32 +196,6 @@ func (sp *SAMLServiceProvider) Validate(el *etree.Element) error {
 	}
 
 	return nil
-
-}
-
-func (sp *SAMLServiceProvider) ValidateEncodedResponse(encodedResponse string) (*etree.Element, error) {
-	raw, err := base64.StdEncoding.DecodeString(encodedResponse)
-	if err != nil {
-		return nil, err
-	}
-
-	doc := etree.NewDocument()
-	err = doc.ReadFromBytes(raw)
-	if err != nil {
-		return nil, err
-	}
-
-	response, err := sp.validationContext().Validate(doc.Root())
-	if err != nil {
-		return nil, err
-	}
-
-	err = sp.Validate(response)
-	if err != nil {
-		return nil, err
-	}
-
-	return response, nil
 }
 
 type ProxyRestriction struct {
@@ -249,137 +222,4 @@ func childPath(space, tag string) string {
 	} else {
 		return "./" + space + ":" + tag
 	}
-}
-
-func (sp *SAMLServiceProvider) VerifyAssertionConditions(assertionElement, conditionsStatement *etree.Element) (*WarningInfo, error) {
-	warningInfo := &WarningInfo{}
-	now := time.Now()
-
-	notBeforeAttr := conditionsStatement.SelectAttr(NotBeforeAttr)
-	if notBeforeAttr != nil {
-		before, err := time.Parse(time.RFC3339, notBeforeAttr.Value)
-		if err != nil {
-			return nil, errors.New("Could not parse 'NotBefore' attribute")
-		}
-
-		if now.Before(before) {
-			warningInfo.InvalidTime = true
-		}
-	}
-	notOnOrAfterAttr := conditionsStatement.SelectAttr(NotOnOrAfterAttr)
-	if notOnOrAfterAttr != nil {
-		after, err := time.Parse(time.RFC3339, notOnOrAfterAttr.Value)
-		if err != nil {
-			return nil, errors.New("Could not parse 'NotOnOrAfter' attribute")
-		}
-		if now.After(after) {
-			warningInfo.InvalidTime = true
-		}
-	}
-
-	audienceRestrictionStatement := conditionsStatement.FindElement(childPath(assertionElement.Space, AudienceRestrictionTag))
-	if audienceRestrictionStatement != nil {
-		audienceStatements := audienceRestrictionStatement.FindElements(childPath(assertionElement.Space, AudienceTag))
-		if len(audienceStatements) == 0 {
-			return nil, errors.New("Missing AudienceStatement")
-		}
-
-		matched := false
-		for _, audienceStatement := range audienceStatements {
-			if audienceStatement.Text() == sp.AudienceURI {
-				matched = true
-			}
-		}
-
-		if !matched {
-			warningInfo.NotInAudience = true
-		}
-	}
-
-	oneTimeUseStatement := conditionsStatement.FindElement(childPath(assertionElement.Space, OneTimeUseTag))
-	if oneTimeUseStatement != nil {
-		warningInfo.OneTimeUse = true
-	}
-
-	proxyRestrictionStatement := conditionsStatement.FindElement(childPath(assertionElement.Space, ProxyRestrictionTag))
-	if proxyRestrictionStatement != nil {
-		proxyRestrictionInfo := &ProxyRestriction{}
-		countAttr := proxyRestrictionStatement.SelectAttr(CountAttr)
-		if countAttr != nil {
-			count, err := strconv.Atoi(countAttr.Value)
-			if err != nil {
-				return nil, errors.New("Could not parse Count attribute")
-			}
-
-			proxyRestrictionInfo.Count = count
-		}
-
-		proxyAudienceStatements := proxyRestrictionStatement.FindElements(childPath(assertionElement.Space, AudienceTag))
-		pas := make([]string, len(proxyAudienceStatements))
-		for i, proxyAudienceStatement := range proxyAudienceStatements {
-			pas[i] = proxyAudienceStatement.Text()
-		}
-
-		proxyRestrictionInfo.Audience = pas
-		warningInfo.ProxyRestriction = proxyRestrictionInfo
-	}
-
-	return warningInfo, nil
-}
-
-func (sp *SAMLServiceProvider) RetrieveAssertionInfo(encodedResponse string) (*AssertionInfo, error) {
-	assertionInfo := &AssertionInfo{}
-
-	el, err := sp.ValidateEncodedResponse(encodedResponse)
-	if err != nil {
-		return nil, err
-	}
-
-	assertionElement := el.FindElement(AssertionTag)
-	if assertionElement == nil {
-		return nil, errors.New("Missing Assertion")
-	}
-
-	//Verify all conditions for the assertion
-	conditionsStatement := assertionElement.FindElement(childPath(assertionElement.Space, ConditionsTag))
-	if conditionsStatement == nil {
-		return nil, errors.New("Missing ConditionsStatement")
-	}
-
-	warningInfo, err := sp.VerifyAssertionConditions(assertionElement, conditionsStatement)
-	if err != nil {
-		return nil, err
-	}
-
-	//Get the NameID
-	subjectStatement := assertionElement.FindElement(childPath(assertionElement.Space, SubjectTag))
-	if subjectStatement == nil {
-		return nil, errors.New("Missing SubjectStatement")
-	}
-
-	nameIdStatement := subjectStatement.FindElement(childPath(assertionElement.Space, NameIdTag))
-	if nameIdStatement == nil {
-		return nil, errors.New("Missing NameIDStatement")
-	}
-	assertionInfo.NameID = nameIdStatement.Text()
-
-	//Get the actual assertion attributes
-	attributeStatement := assertionElement.FindElement(childPath(assertionElement.Space, AttributeStatementTag))
-	if attributeStatement == nil {
-		return nil, errors.New("Missing AttributeStatement")
-	}
-
-	info := make(map[string]string)
-	for _, child := range attributeStatement.ChildElements() {
-		nameAttr := child.SelectAttr(NameAttr)
-		attributeValue := child.FindElement(childPath(child.Space, AttributeValueTag))
-		if attributeValue == nil {
-			return nil, errors.New("Missing AttributeValue")
-		}
-		info[nameAttr.Value] = attributeValue.Text()
-	}
-
-	assertionInfo.Values = info
-	assertionInfo.WarningInfo = warningInfo
-	return assertionInfo, nil
 }
